+++ src/Net.Sf.Dbdeploy/Resources/CreateSchemaVersionTable.mssql.sql
@@ -6,7 +6,7 @@ END
 CREATE TABLE $(QualifiedTableName) (
 	ChangeId INT IDENTITY(1,1) NOT NULL,
 	Folder VARCHAR(256) NOT NULL,
-	ScriptNumber SMALLINT NOT NULL,
+	ScriptNumber BIGINT NOT NULL,
 	ScriptName VARCHAR(512) NOT NULL,
 	StartDate DATETIME NOT NULL,
 	CompleteDate DATETIME NULL,
